+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.2.0] - 2026-01-28
+
+### Added
+
+- Transaction model with description, amount, category, date, and status
+- Account model with balance rollover logic ($Amount\_owed - Amount\_paid = New\ Owed$).
+- Pytest configuration for automated test discovery.
+
 ## [0.1.0] - 2026-01-28
 
 - Initialized project structure and environment.